// hq-app/containers/Stores/types.ts
/* eslint-disable @typescript-eslint/ban-types */
/* eslint-disable @typescript-eslint/no-explicit-any */
export type ADD_UPDATE_FROM_PROPS = {
  storePaymentMethods?: any;
  countryNameId?: number | string;
  cityNameId?: number | string;
  stateNameId?: number | string;
  id?: number;
  store?: object;
  country?: object;
  state?: object;
  city?: object;
  productsData?: any;
  currencyId?: string | number | null;
  currencyName?: string;
  paymentMethodName?: any;
  currency?: { name?: string; id?: string };
  storeTaxes?: any;
  floor?: string | number;
  address?: string;
  contactNumber?: string;
  email?: string;
  countryName?: string;
  countryId?: number;
  stateName?: string;
  stateId?: number;
  cityName?: string;
  cityId?: number;
  entityId?: number;
  numberSequence?: string;
  tax?: Array<any> | undefined;
  name?: string | undefined;
  code?: string | undefined;
  storeId?: string;
  storeName?: string;
  storeType?: string | number;
  paymentMethod?: string | number;
  todayFilterType?: string | number;
  type?: string | number | undefined;
  active?: boolean;
  priceExclusiveTax?: boolean;
  parkEnabled?: boolean;
  discountExclusiveTax?: boolean;
  storeLogo?: any;
  showMenuImages?: boolean;
  addresses?: any | undefined;
  assignedProducts?: any | undefined;
  AssignedProducts?: any | undefined;
  storeAddress?: object | undefined | any;
  storeDetails?: STORE_DETAILS;
  callBack?: () => void;
  ModifiedTax?: any;
  ModifiedAddress?: any;
  deliveryCharges?: number | string;
};

export type STORE_DETAILS = {
  hide?: boolean;
  id?: number;
  activeTab?: string;
  code?: string;
  name?: string;
  type?: string | number;
  paymentMethod?: string | number;
  todayFilterType?: string | number;
  active?: Boolean;
  priceExclusiveTax?: Boolean;
  parkEnabled?: Boolean;
  discountExclusiveTax?: Boolean;
  showMenuImages?: Boolean;
  currency?: any;
  floor?: string | number;
  tax?: any;
  numberSequence?: string;
  storeAddress?: any | undefined;
  storeProducts?: Array<any> | undefined | any;
  storeTaxes?: Array<any> | undefined;
  data?: Array<any> | undefined | any;
  deliveryCharges?: string | number;
  storeLogo?: string;
  storePaymentMethods?: any;
};